move the repo to use import maps
