+++ .github/workflows/ci.yml
@@ -75,12 +75,16 @@ jobs:
 
       - name: Type check website
         run: deno check --remote www/main.ts
+        working-directory: www/
 
       - name: Type check examples/counter
         run: deno check --remote examples/counter/main.ts
+        working-directory: examples/counter
 
       - name: Type check tests/fixture
         run: deno check tests/fixture/main.ts
+        working-directory: tests/fixture/
 
       - name: Type check tests/fixture_error
-        run: deno check tests/fixture_error/main.ts
+        run: deno check main.ts
+        working-directory: tests/fixture_error/
