--- examples/counter/deps.client.ts
@@ -1,1 +1,0 @@
-export * from "../../runtime.ts";
