--- tests/fixture/deps.client.ts
@@ -1,7 +1,0 @@
-/// <reference no-default-lib="true" />
-/// <reference lib="dom" />
-/// <reference lib="dom.asynciterable" />
-/// <reference lib="deno.ns" />
-/// <reference lib="deno.unstable" />
-
-export * from "../../runtime.ts";
