--- tests/fixture/deps.server.ts
@@ -1,1 +1,0 @@
-export * from "../../server.ts";
