+++ tests/fixture/main.ts
@@ -1,4 +1,4 @@
-import { start } from "./deps.server.ts";
+import { start } from "$fresh/server.ts";
 import routes from "./fresh.gen.ts";
 
 await start(routes);
