+++ tests/fixture/routes/_404.tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h, UnknownPageProps } from "../deps.client.ts";
+import { h, UnknownPageProps } from "$fresh/runtime.ts";
 
 export default function NotFoundPage({ url }: UnknownPageProps) {
   return <p>404 not found: {url.pathname}</p>;
