+++ tests/fixture/routes/_app.tsx
@@ -1,7 +1,7 @@
 /** @jsx h */
 /** @jsxFrag Fragment */
 
-import { AppProps, Fragment, h, Head } from "../deps.client.ts";
+import { AppProps, Fragment, h, Head } from "$fresh/runtime.ts";
 
 export default function App(props: AppProps) {
   return (
