+++ tests/fixture/routes/_middleware.ts
@@ -1,4 +1,4 @@
-import { MiddlewareHandlerContext } from "../deps.server.ts";
+import { MiddlewareHandlerContext } from "$fresh/server.ts";
 
 export async function handler(_req: Request, ctx: MiddlewareHandlerContext) {
   ctx.state.root = "root_mw";
