+++ tests/fixture/routes/assetsCaching/index.tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { asset, h } from "../../deps.client.ts";
+import { asset, h } from "$fresh/runtime.ts";
 import Test from "../../islands/Test.tsx";
 
 export default function Home() {
