+++ tests/fixture/routes/props/[id].tsx
@@ -1,6 +1,6 @@
 /** @jsx h */
 
-import { h, PageProps } from "../../deps.client.ts";
+import { h, PageProps } from "$fresh/runtime.ts";
 
 export default function Home(props: PageProps) {
   return <div>{JSON.stringify(props)}</div>;
