--- tests/fixture_error/deps.client.ts
@@ -1,1 +1,0 @@
-export * from "../../runtime.ts";
