--- tests/fixture_error/deps.server.ts
@@ -1,1 +1,0 @@
-export * from "../../server.ts";
