+++ tests/fixture_error/import_map.json
@@ -1,0 +1,5 @@
+{
+  "imports": {
+    "$fresh/": "../../"
+  }
+}
