+++ tests/fixture_error/routes/index.tsx
@@ -1,5 +1,5 @@
 /** @jsx h */
-import { h } from "../deps.client.ts";
+import { h } from "$fresh/runtime.ts";
 
 export default function Home(): null {
   throw new Error("boom!");
