+++ www/routes/gfm.css.ts
@@ -1,4 +1,5 @@
-import { gfm, Handlers } from "../deps.server.ts";
+import { Handlers } from "$fresh/server.ts";
+import { gfm } from "../utils/markdown.ts";
 
 // TODO(lucacasonato): hash the file and use the hash as the filename, and serve
 // with high-cacheability headers.
