+++ www/utils/markdown.ts
@@ -1,7 +1,3 @@
-export * from "../server.ts";
-
-export { virtualSheet } from "https://esm.sh/twind@0.16.16/sheets";
-
 export * as gfm from "https://deno.land/x/gfm@0.1.19/mod.ts";
 import "https://esm.sh/prismjs@1.27.0/components/prism-jsx.js?no-check";
 import "https://esm.sh/prismjs@1.27.0/components/prism-typescript.js?no-check";
