+++ www/utils/twind.ts
@@ -1,7 +1,6 @@
-export * from "../runtime.ts";
-import { IS_BROWSER } from "../runtime.ts";
-import { apply, setup, tw } from "https://esm.sh/twind@0.16.16";
-import * as colors from "https://esm.sh/twind@0.16.16/colors";
+import { IS_BROWSER } from "$fresh/runtime.ts";
+import { apply, setup, tw } from "$twind";
+import * as colors from "$twind/colors";
 export { apply, setup, tw };
 export const theme = {
   colors: {
